intermediate state
