+++ Pages/dashboard.html
@@ -2,7 +2,7 @@
 <html>
 <head>
   <title>Dashboard</title>
-  <link rel="stylesheet" href="style.css">
+  <link rel="stylesheet" href="../Styles/user-account-style.css">
 </head>
 <body>
 
@@ -14,7 +14,7 @@
   <button onclick="logout()">Logout</button>
 </div>
 
-<script type="module" src="auth.js"></script>
+<script type="module" src="../Scripts/great_outdoors.js"></script>
 
 </body>
 </html>
