+++ Pages/login.html
@@ -2,7 +2,7 @@
 <html>
 <head>
   <title>Login</title>
-  <link rel="stylesheet" href="style.css">
+  <link rel="stylesheet" href="../Styles/user-account-style.css">
 </head>
 <body>
 
@@ -20,7 +20,7 @@
   </p>
 </div>
 
-<script type="module" src="auth.js"></script>
+<script type="module" src="../Scripts/great_outdoors.js"></script>
 
 </body>
 </html>
