+++ Scripts/great_outdoors.js
@@ -15,6 +15,14 @@ import {
   serverTimestamp,
 } from "https://www.gstatic.com/firebasejs/10.12.2/firebase-firestore.js";
 
+import {
+  getAuth,
+  createUserWithEmailAndPassword,
+  signInWithEmailAndPassword,
+  signOut,
+  onAuthStateChanged
+} from "https://www.gstatic.com/firebasejs/10.12.2/firebase-auth.js";
+
 // ── Re-use your existing firebaseConfig ──────────────────────
 const firebaseConfig = {
   apiKey: "AIzaSyABZKs8MP6NwOjgnKB7qNdh11FOftnDxCk",
@@ -261,4 +269,74 @@ export async function updateTrail(id, partialData) {
 export async function deleteTrail(id) {
   await deleteDoc(doc(db, "trails", id));
   console.log("Trail deleted:", id);
+}
+
+
+
+// AUTH
+// SIGN UP
+window.signUp = function () {
+
+  const email = document.getElementById("signup-email").value;
+  const password = document.getElementById("signup-password").value;
+
+  createUserWithEmailAndPassword(auth, email, password)
+    .then(() => {
+      alert("Account Created!");
+      window.location.href = "dashboard.html";
+    })
+    .catch((error) => {
+      alert(error.message);
+    });
+};
+
+
+// LOGIN
+window.login = function () {
+
+  const email = document.getElementById("login-email").value;
+  const password = document.getElementById("login-password").value;
+
+  signInWithEmailAndPassword(auth, email, password)
+    .then(() => {
+      alert("Logged In!");
+      window.location.href = "dashboard.html";
+    })
+    .catch((error) => {
+      alert(error.message);
+    });
+};
+
+
+// LOGOUT
+window.logout = function () {
+
+  signOut(auth)
+    .then(() => {
+      alert("Logged Out");
+      window.location.href = "login.html";
+    });
+};
+
+
+// Protect Dashboard
+onAuthStateChanged(auth, (user) => {
+
+  if (
+    window.location.pathname.includes("dashboard.html")
+    && !user
+  ) {
+    window.location.href = "login.html";
+  }
+});
+
+
+
+
+function showAlt(x) {
+  document.getElementById("alttext").innerHTML=x.alt;
+}
+
+function hideAlt(x) {
+  document.getElementById("alttext").innerHTML="";
 }
