--- scripts/myscripts.js
@@ -1,8 +1,0 @@
-
-function showAlt(x) {
-  document.getElementById("alttext").innerHTML=x.alt;
-}
-
-function hideAlt(x) {
-  document.getElementById("alttext").innerHTML="";
-}
